Fix reflow computations which were causing widget overlap
