+++ app/client/src/utils/autoHeight/constants.ts
@@ -6,3 +6,13 @@ export type TreeNode = {
   originalTopRow: number;
   originalBottomRow: number;
 };
+
+export type NodeSpace = {
+  left: number;
+  right: number;
+  top: number;
+  bottom: number;
+  id: string;
+};
+
+export const MAX_BOX_SIZE = 20000;
